requirements.txt
fastapi
sqlalchemy
redis
celery
aiohttp
uvicorn
python-dotenv